Revert "LPS-174464 Move back as inner classes."
This reverts commit 2e7cd234107dc8e3e6395104ef24d6b7af23cacb.
